% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{gcheatmap}
\alias{gcheatmap}
\title{Heatmap of genetic correlations}
\usage{
gcheatmap(data = gcatlas, distfun = cor2dist, symm = TRUE,
  margins = c(9, 9), fix = TRUE, ...)
}
\arguments{
\item{data}{An object of the same structure as \code{gcatlas}}

\item{distfun}{The function to convert a correlation matrix to a 
distance matrix for generating the dendrograms}

\item{symm}{Logical switch indicating a heatmap of a symmetric matrix; 
passed on to \code{heatmap} and probably best left unchanged}

\item{margins}{Numerical vector of length two, indicating how much 
space is reserved for the row/column labels. This is the same
argument as for \code{heatmap}, but with a different default
so that the trait names are reasonably readable with normal
device size}

\item{fix}{Logical expression indicating whether correlations outside 
of [-1,1] should be set to the limits}

\item{...}{Extra arguments passed on to \code{heatmap}}
}
\value{
Invisibly, the same as function \code{heatmap}. This function
        is generally invoked to generate a plot.
}
\description{
Plot a heatmap of genetic correlations
}
\examples{
## Default heatmap of the full atlas
gcheatmap()	
}
\seealso{
\code{\link{heatmap}}, \code{\link{cor2dist}}, \code{\link{gcatlas}}
}

