% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{gcmatrix}
\alias{gcmatrix}
\title{Convert Atlas of Genetic Correlations to a matrix}
\usage{
gcmatrix(data = gcatlas, type = c("rg", "se", "z", "p"), diag_value)
}
\arguments{
\item{data}{The object from which to extract the variable - by default, 
the full \code{gcatlas} object, but a valid subset will 
work, too}

\item{type}{The variable to extract, one of \code{rg} (default), \code{se}
\code{z} and \code{p}}

\item{diag_value}{An optional value for the diagonal of the matrix; 
if missing, the function will use a reasonable 
default, based on the value of \code{type}}
}
\value{
A square numerical matrix
}
\description{
This function extracts any of the variables in the \code{gcatlas} 
object and returns it as a square matrix, with traits as rows and columns.
}
\seealso{
\code{\link{gcatlas}}
}

