% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\docType{package}
\name{GeneCorrAtlas}
\alias{GeneCorrAtlas}
\alias{GeneCorrAtlas-package}
\title{Display the Atlas of Genetic Correlations}
\source{
\url{http://www.nature.com/ng/journal/vaop/ncurrent/extref/ng.3406-S2.csv}
}
\description{
Bulik-Sullivan et al. recently (2105) published an atlas of genetic
correlations for 50 different traits. This package wraps the more than 1000
correlation coefficients involved and offers some convenient functions to
extract and display the data
}
\references{
Bulik-Sullivan B, Finucane HK, Anttila V, Gusev A, Day FR, Loh P-R,
   et al. An atlas of genetic correlations across human diseases and traits.
   Nat Genet [Internet]. 2015 Sep 28 [cited 2015 Oct 5];advance online publication
}

