% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{match_trait_names}
\alias{match_trait_names}
\title{Match trait names to a data set}
\usage{
match_trait_names(x, data = gcatlas, unique = FALSE)
}
\arguments{
\item{x}{a vector of (possibly abbreviated) trait names}

\item{data}{the data frame against which the trait names are compared}

\item{unique}{Logical switch indicating whether a unique match is
required; if \code{TRUE}, thorws an error if \code{x}
cannot be matched, or matches more than one trait}
}
\description{
This is a utility function that takes a vector of possibly abbreviated trait
names and returns the matching trait names from a data frame with the same
format as \code{\link{gcatlas}} (which is the default).
}
\seealso{
\code{\link{gcatlas}}
}

