% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{plotPairedTraits}
\alias{plotPairedTraits}
\title{Scatterplot of genetic correlations for two traits}
\usage{
plotPairedTraits(xtr, ytr, data = gcatlas, co = NA)
}
\arguments{
\item{xtr, ytr}{The names of the traits to be displayed on the x- and 
y-axis; can be abbreviated, but must be unique.}

\item{data}{Either the full atlas of genetic correlations (\code{gcatlas},
the default), or a valid subset, from which to draw correlations.}

\item{co}{A numerical cutoff below which p-values are considered 
statistically significant. \code{NA} (the default) indicates
that no highlighting should take place.}
}
\value{
The plot as an ggplot2 object. Note that the plotting data 
        can be easily extracted for further processing, see Examples.
}
\description{
Given the atlas of genetic correlations or a valid subset, plus two trait
names, this function produces a scatterplot of the genetic correlations
of one trait against the genetic correlations of the other. Correlations
that are statistically significant at a given level can be highlighted.
}
\examples{
## Default invocation: no highlighting
plotPairedTraits("ADHD", "Alz")

## Correlations with p-values less than 0.01 are highlighted
plotPairedTraits("ADHD", "Alz", co=0.01)

## Extract the data
x = plotPairedTraits("Cor", "T2D", co=0.05)
x$data
## Useful extra attributes
attributes(x$data)
}
\seealso{
\code{\link{gcatlas}}, \code{\link{sel_trt}}
}

