% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{plotTrait}
\alias{plotTrait}
\title{Plot all correlations with a specific trait}
\usage{
plotTrait(trait, data = gcatlas, conf = 0.95, sort = TRUE,
  rlim = c(-0.5, 1))
}
\arguments{
\item{trait}{The name of the trait for which to plot the correlations;
can be abbreviated, but must be unique.}

\item{data}{Either the full atlas of genetic correlations (\code{gcatlas},
the default), or a valid subset, from which to draw correlations.}

\item{conf}{The confidence level for the confidence intervals}

\item{sort}{Logical switch indicating whether to sort the traits by 
by correlation (default: \code{TRUE})}

\item{rlim}{Cutoff values for plotting the confidence limits, so that
wide confidence limits do not dominate the plot}
}
\value{
The plot as an ggplot2 object. Note that the plotting data 
        can be easily extracted for further processing, see Examples.
}
\description{
Given the atlas of genetic correlations or a valid subset, plus a trait
name, this function plots all correlations of the specified trait and 
all other traits in the correlation set. The plotting data (including
confidence intervals) is returned invisibly.
}
\examples{
## Default invocation
plotTrait("ADHD")

## Subset of the full atlas
plotTrait("Alz", sel_trt("Obesity"))

## Extract the plotting data
x = plotTrait("Cor")
x$data
## Useful extra attributes
attributes(x$data)
}
\seealso{
\code{\link{gcatlas}}, \code{\link{sel_trt}}
}

