% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{read_gcatlas}
\alias{read_gcatlas}
\title{Read Atlas of Genetic Correlations}
\source{
\url{http://www.nature.com/ng/journal/vaop/ncurrent/extref/ng.3406-S2.csv}
}
\usage{
read_gcatlas(x)
}
\arguments{
\item{x}{name of the supplementary file (by default, a link to the original
   publication)}
}
\value{
The data frame described at \code{\link{gcatlas}}
}
\description{
Read the supplementary file from disk or from the web. This is the function
originally used to generate the pre-packaged data object \code{\link{gcatlas}}.
Only use this function if you want to roll your own.
}
\references{
Bulik-Sullivan B, Finucane HK, Anttila V, Gusev A, Day FR, Loh P-R,
   et al. An atlas of genetic correlations across human diseases and traits.
   Nat Genet [Internet]. 2015 Sep 28 [cited 2015 Oct 5];advance online publication
}

