% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{sel_trt}
\alias{sel_trt}
\title{Select traits from the full atlas}
\usage{
sel_trt(..., data = gcatlas, drop = FALSE)
}
\arguments{
\item{...}{a list of trait names (possibly abbreviated)}

\item{data}{the name of the object to subset; by default the full data set}

\item{drop}{logical expression indicating whether to return the specified traits
(\code{FALSE}, default) or to exclude them from the data (\code{TRUE}).}
}
\value{
An object of the same structure as \code{gcatlas}
}
\description{
Subsets a data frame with the same format as \code{\link{gcatlas}} to
include only traits from a specified list of trait names. These names can be
abbreviated.
}
\examples{
x = sel_trt("ADHD", "Alz", "T2D")
x
gcmatrix(x)
}
\seealso{
\code{\link{gcatlas}}
}

