% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{traits}
\alias{traits}
\title{Extract trait names from the atlas of genetic correlations}
\usage{
traits(data = gcatlas)
}
\arguments{
\item{data}{the object from which to extract trait names (default: 
the full atlas)}
}
\value{
A character vector of trait names, in the order they appear
        in \code{Trait1}/\code{Trait2} in \code{data}. This is \emph{not}
        sorted, only almost sorted
}
\description{
A simple helper function that returns a sorted list of traits found
in a \code{gcatlas}-type of object
}
\examples{
traits()
}
\seealso{
\code{\link{gcatlas}}
}

