% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GeneCorrAtlas.R
\name{sel_trt}
\alias{sel_trt}
\title{Select traits from the full atlas}
\usage{
sel_trt(x, data = gcatlas, drop = FALSE)
}
\arguments{
\item{x}{a vector of trait names (possibly abbreviated)}

\item{data}{the name of the object to subset; by default the full data set}

\item{drop}{logical expression indicating whether to return the specified traits
(\code{FALSE}, default) or to exlclude them from the data (\code{TRUE}).}
}
\description{
Subsets the a data frame with the same format as \code{\link{gcatlas}} to
include only traits from a specified list of trait names. These names can be
abbreviated.
}
\seealso{
\code{\link{gcatlas}}
}

